Fix parsing of non-integer values

+++ src/lib/keyframe.ts
@@ -40,7 +40,7 @@ function parseKeyframeEntries(value: string): Array<KeyFrameEntry> {
       values: values
         .split(" ")
         .map((x) => x.trim())
-        .map((y) => Number.parseInt(y)),
+        .map((y) => Number.parseFloat(y)),
     };
   };
